* upstream/main:
  Check documentation labels, fixes #627 (#671)
  Fix integer overflow in SortedRanges.invertOnNew.  Fixes #666. (#667)
  Fix integer overflow in SortedRanges.subRangeByPos. Fixes #664 (#665)
  Deploy example scripts and data (#562)
  port the numba integration prototype from Enterprise to OSS, fixes feature-523 (#638)
  Fix IllegalStateException in TwoValuesContainer triggered from Index.insert(Chunk). Fixes #652. (#653)
  Javadoc search (#645)
  Support for jpy integration junit tests (#632)
  Populate snapshot request in js api, handle optional left table (#648)
  Fix a link formatting bug in CONTRIBUTING.md (#629)

+++ DB/src/main/java/io/deephaven/db/v2/utils/sortedranges/SortedRanges.java
@@ -1040,7 +1040,7 @@ public abstract class SortedRanges extends RefCountedCow<SortedRanges> implement
         // We don't want to do two passes, we allocated an array big enough instead.
         final boolean brokenInitialRange = startPos < pos;
         int ansLen = count - i + (brokenInitialRange ? 2 : 1);
-        ansLen = Math.min(ansLen, (int) (inputRangeSpan + 1));
+        ansLen = (int) Math.min(ansLen, (inputRangeSpan + 1));
         final SortedRanges ans = makeMyTypeAndOffset(ansLen);
         ans.count = 0;
         ans.cardinality = 0;
@@ -2191,7 +2191,7 @@ public abstract class SortedRanges extends RefCountedCow<SortedRanges> implement
     public final TreeIndexImpl invertRangeOnNew(final long start, final long end, final long maxPosition) {
         final long packedStart = pack(start);
         int i = 0;
-        int pos = 0;
+        long pos = 0;
         long data = packedGet(i);
         boolean neg = false;
         long pendingStart = -1;
@@ -2260,7 +2260,7 @@ public abstract class SortedRanges extends RefCountedCow<SortedRanges> implement
         long end = rit.currentRangeEnd();
         long packedStart = pack(start);
         int i = 0;
-        int pos = 0;
+        long pos = 0;
         long data = packedGet(i);
         boolean neg = false;
         long pendingStart = -1;
